Align alpha type with the SP interface

+++ include/onmt/SentencePiece.h
@@ -14,7 +14,7 @@ namespace onmt
   public:
     SentencePiece(const std::string& model_path);
 
-    void enable_regularization(int nbest_size, double alpha);
+    void enable_regularization(int nbest_size, float alpha);
 
     std::vector<std::string> encode(const std::string& str) const override;
     std::vector<AnnotatedToken> encode_and_annotate(const AnnotatedToken& token) const override;
@@ -22,7 +22,7 @@ namespace onmt
   private:
     sentencepiece::SentencePieceProcessor _processor;
     int _nbest_size;
-    double _alpha;
+    float _alpha;
   };
 
 }
